pyproject.toml
[tool.poetry]
name = "short-movie"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.9.2"
openai = "^1.47.1"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.2"
litellm = "^1.48.1"
loguru = "^0.7.2"
pylint = "^3.3.1"
isort = "^5.13.2"
instructor = "^1.4.3"
jinja2 = "^3.1.4"
google-api-python-client = "^2.147.0"
numpy = "^2.1.1"
requests = "^2.32.3"
streamlit = "^1.39.0"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
